requirements.txt
h2o
scikit-learn==1.1.1
pyyaml
pandas
pyarrow
numpy==1.24.4